refactor: keep HTTP admission policy in Work transport
